Update build scripts

+++ windows/build.sh
@@ -4,12 +4,15 @@
 
 APP=hello-world-gtk
 
-echo "Setting up virtual environment..."
-
-python3 -m venv --system-site-packages venv
-. venv/bin/activate
-python3 -m pip install --upgrade pip
-PYINSTALLER_COMPILE_BOOTLOADER=1 PYI_STATIC_ZLIB=1 python3 -m pip install -r ../requirements.txt
+if [ ! -d ../venv ]; then
+	echo "Setting up virtual environment..."
+	python3 -m venv --system-site-packages venv
+	. venv/bin/activate
+	python3 -m pip install --upgrade pip
+	PYINSTALLER_COMPILE_BOOTLOADER=1 PYI_STATIC_ZLIB=1 python3 -m pip install -r ../requirements.txt
+else
+	. ../venv/bin/activate
+fi
 
 echo "Running pyinstaller..."
 
@@ -39,4 +42,6 @@ rm INSTALLSIZE
 rm -r build
 rm -r dist/*/*
 rm -r dist
-rm -r venv
+if [ ! -d ../venv ]; then
+	rm -r venv
+fi
